Added new module.
[UPD] Updated code with working items.

[UPD] Improved logic and added config component domain

[UPD] Added readme and default qty

[UPD] pre-commit

[UPD] pre-commit

[UPD] pre-commit

Allow search config component by internal ref

Update product_configurator_mrp_component/models/mrp_bom_line_config.py

pre-commit

[ADD] Added filter

change configurable component availability

Changes the logic of the configurable components domain. It now first checks that the child has all the attributes on the parent, and if so it then makes sure that the child has all the values for that line in the parent. This should allow a child which can have more values then the parent to still be added as a configurable component.

Co-Authored-By: Daniel Reis <dreis@opensourceintegrators.com>