[14.0][IMP] product_configurator_mrp (Set BoM Sequences)
Sets sequences on the BoM so proper BoM is used on the MO.
When the variant is created by the configurator, so is it's BoM however the BoM sequences typically aren't set. When Odoo creates an MO, it looks through the BoM's of the related product template, and uses the first one it finds that matches the product or a BoM that doesn't have the product variant.
If the main BoM (aka Master BoM) found first, it will be used instead of the variant's BoM since it doesn't have a variant applied to it.
Setting the sequences, and making the BoM's without variants have a high sequence resolves this issue.