[FIX][14.0] product_configurator (product.template.attribute.value Inactive)
We found in a certain scenario, if a attribute/value is added twice to a product, then one is removed, the variant that gets created will contain both values. This is because the product.template.attribute.value doesn't have a normal 'active' field, instead has ptav_active so the search returns active and inactive values. This small code change fixes this issue.